feat: add troubleshooting documentation for health probes and Arcjet
I wasn't able to find anywhere else that included a copy-and-paste
example of a middleware with Arcjet, so I included the fix for Next.js.

Please let me know if this should be moved somewhere else, I wasn't
able to find a good home for this in the existing ontology so I took
a guess.

Signed-off-by: Xe Iaso <me@xeiaso.net>