chore: improve adding new frameworks to `lib/prefs`
The first step when adding a new framework is to add it to
`src/lib/prefs.ts`.
Previously new values had to be added in several places,
and the types for these different places were not tied
together.
This commit ties these types together and reduces the needed
changes to one line.