docs: add docs on Client IP
This is based on current workings, with additions of how it *could* work,
marked with visible TODOs, linking to issues in `arcjet/arcjet-js`.

This introduces a new “Concepts” section.
Some users want to get started, some want to see Arcjets features,
but some come in with a problem around a particular “topic”.
I hope to put stuff like “Environment variables”, “Request body”,
“Fingerprints”, and such typically finicky stuff under here.
I explicitly want these a bit higher that “Advanced”, to not dissuade
folks.
But that’s where I want to go and as this is the lonely first, open to
have it somewhere else.

(note: not draft because I explicitly want to hear thoughts)

Closes GH-703.