Add Intro pages for features
This PR adds intro pages for each rule/primitive/product.
It moves existing sections, mostly from “Concepts”, into these pages.
Fills in gaps. And streamlines the result.

The purpose of this page is to answer questions that users have before
they get started using code or delve deeply into the reference.
Each page gives a short overview, explains when to use this and when to
use something else, how it works (what happens locally, what in the
Cloud API), and whether it is available on different plans.

A future PR can move the remaining “Concepts” sections, probably also
the one-off pages such as “Identifying bots”, “Configuration” (under
Rate limiting), into “References”.
Another PR can link to these from the docs homepage.

This is a step in the direction of improving the maintenance multiplier
of `feature * framework (* language)`.

References ENG-133.